feat: changes catalog and OCP release heads only workflows to use implicit ranges
To remove the requirement to merge FBCs, packages ranges are being used
in metadata for heads only workflows to remove generate FBCs that are scoped
per the configuration instead of generating differential FBCs and merging with previously
generated FBCs. The OCP releases heads only workflow has been updated work in the same way.

BREAKING CHANGE: This change adds OCPMetadata and modifies OperatorMetadata on the Metadata Spec.
It also removes the catalog merging workflow.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>